# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Admission Application API",
        "description": "REST API for the Secondary School Admission Registration System. Manages applicant biodata, next-of-kin information, and health records.",
        "contact": {
            "email": "admin@school.edu.ng"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Version 1"
        }
    ],
    "paths": {
        "/applicant/auth/login": {
            "post": {
                "tags": [
                    "Applicant Auth"
                ],
                "summary": "Applicant login",
                "description": "Authenticates an applicant using their application number and password. Returns a Bearer token for use on all protected applicant endpoints.",
                "operationId": "7f418eb0a1b894e3454a99b69e07da1b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "chukwuemeka@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "accessToken": {
                                            "type": "string",
                                            "example": "2|abc123token..."
                                        },
                                        "applicant": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "application_number": {
                                                    "type": "string",
                                                    "example": "APP-6A36A61045B86"
                                                },
                                                "first_name": {
                                                    "type": "string",
                                                    "example": "Chukwuemeka"
                                                },
                                                "last_name": {
                                                    "type": "string",
                                                    "example": "Okafor"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "chukwuemeka@example.com"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/applicant/auth/logout": {
            "post": {
                "tags": [
                    "Applicant Auth"
                ],
                "summary": "Applicant logout",
                "description": "Revokes the current applicant Bearer token.",
                "operationId": "9ed2cd2d695fb88f2d3a023e33f11502",
                "responses": {
                    "200": {
                        "description": "Logged out",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden — token does not belong to an applicant"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/applicant/auth/me": {
            "get": {
                "tags": [
                    "Applicant Auth"
                ],
                "summary": "Get the authenticated applicant profile",
                "description": "Returns basic profile information for the currently authenticated applicant.",
                "operationId": "0feb0aab155f327c26707a3a3e078432",
                "responses": {
                    "200": {
                        "description": "Applicant profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "application_number": {
                                            "type": "string",
                                            "example": "APP-6A36A61045B86"
                                        },
                                        "first_name": {
                                            "type": "string",
                                            "example": "Chukwuemeka"
                                        },
                                        "last_name": {
                                            "type": "string",
                                            "example": "Okafor"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "chukwuemeka@example.com"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "pending"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden — token does not belong to an applicant"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter the token returned by POST /api/v1/auth/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}